requirements.txt
azure-cognitiveservices-speech>=1.32.0
openai>=1.13.0
openai[realtime]
gradio>=4.0.0
matplotlib>=3.5.0
pandas>=2.0.0
plotly>=5.0.0
numpy>=1.23.0
azure-identity>=1.15.0
python-dotenv>=1.0.0
pydub
aiohttp